refactor: rename motions.ts to calendarMotions.ts
